Use duck typing instead of a type check.
StringIO isn't an IO, and so can't be used to generate a "data" type. Instead
of testing an object's type, the code is only interested if that object has a
read method, which StringIO instances do.

It's possible this may break existing code, but arguably if someone wants a
plist to contain a String, they should be calling "string" on their StringIO
instances to encode real strings.