When possible, pack the entire offsets table at once.
Since the offsets table is already in an array, it's more efficient to pack
all of the offsets at once. As long as offsets are 4 bytes or smaller, this
is easy because Array#pack can pack a 32-bit integer in network byte order.
64-bit integers must be handled one-by-one because pack doesn't have an
operator for them that converts from host to network (big endian) byte order.

For an array with 1000 entries, each a hash containing six entries (around
12000 objects in total), this is about a 12% performance improvement.