Refactor building unique object table.
unique_and_count_values actually did a lot of the same work as the code
that emits the binary encoding, especially for property lists that made
heavy use of strings (computing lengths and character encoding conversions,
mostly). Worse, even though it built a unique table, it was thrown away,
only to built again when the actual binary plist data was generated.

I've refactored this by removing unique_and_count_values and implementing a
new, simple count_object_refs method. At this stage, the only thing we need
to know is how many object refs there are, which boils down to the number of
array elements plus the number of hash keys and values. The reason is that
when generating the binary data for those types, we need to know how big the
resulting integer indices are (1, 2, 4 or possibly 8 bytes, though only
theoretically since that means more than 4 billions object refs and a
ridiculously big plist).

The other work that unique_and_count_values did was to count the number of
bytes in the object table, so it could compute the values needed in the
plist trailing. I moved these calculations into the various *_to_binary
methods.

All together, this results in about a 23% performance improvement on my test
data.