Amend prior fix to Binary.ascii_string?.
String#scan is measurably faster on Ruby 1.8 than walking through the bytes,
so I fixed the SyntaxError on Ruby 1.9 in the regular expression without
breaking it in Ruby 1.8. Also, Ruby 1.9 has String#ascii_only? that performs
this very same test, so it makes sense to use it as it's likely to be the
faster choice there.