Fix for native_types returning Blob data
Previously, native_types would return an object's `decoded_value` if
it was determined that the object was a CFData type.
Because the decoded_value was a String, the native type returned
would contain String data instead of CFPropertyList::Blob data. This caused
a problem when you tried to convert that native type to a plist - the underlying
parser would raise an Encoding error whenever it tried to handle binary data
that was in a String format instead of a CFData format.

This commit returns the `decoded_value` wrapped in a CFPropertyList::Blob
declaration, which will return a native type containing an item of
CFPropertyList::Blob, instead of type String, which CFPropertyList correctly
handles when you try to convert it to a plist.