Move CFPropertyList to a more idomatic layout
Currently if a user were to want to use CFPropertyList without installing it
via rubygems, the most common installation would have been into site_ruby. This
would result in a large number of files in site_ruby for a single project,
which is unusual. Ruby projects generally have a single ruby file in
site_ruby/vendor_ruby and a directory in site_ruby/vendor_ruby with the
remainder of the classes and modules. This commit updates CFPropertyList to
follow that structure, with cfpropertylist.rb under lib and the remainder of
the classes and modules under lib/cfpropertylist. The entry point to the
library remains as easy as `require 'cfpropertylist'`.