Update tests for new lib structure
The previous commit updated the lib structure to move most of the library under
lib/cfpropertylist. This means that the requires in the tests need to be
updated to find the files correctly. The updated requires presume that the lib
directory of cfpropertylist is in the load path of the ruby running the test
suite.