begin to implement uid object type (not yet finished)
Apple has a special object type uid. This basically is an integer with a
special object type code. In XML this is represented via
<dict><key>CF$UID</key><integer>...</integer></dict> (yes, via a
dictionary with just one only value - WTF?).

To match XML behavior I currently create a dictionary like described
above in binary format as well. But I'm not sure if this is the way we
should go…