Changes "native format" to "little endian" format (fixes #46).
So far, the "native format" for reading real numbers and dates was used.
That meant that the test suite, which is reading pregenerated .plist
files, was working on LE platforms, but was failing on BE platforms
(ppc64). This patch make the test suite pass on all platforms by
explicitly spefiying the "little endian" format of real numbers and
dates fields.